+++ src/main/java/com/arcanc/biomorphosis/content/block/block_entity/ber/BioForgeRenderer.java
@@ -12,7 +12,7 @@ package com.arcanc.biomorphosis.content.block.block_entity.ber;
 import com.arcanc.biomorphosis.content.block.BioForgeBlock;
 import com.arcanc.biomorphosis.content.block.block_entity.BioForge;
 import com.arcanc.biomorphosis.util.Database;
-import com.arcanc.pulselib.content.event.CustomEvents;
+import com.arcanc.pulselib.content.event.PulseLibEvents;
 import com.arcanc.pulselib.content.renderer.PBlockRenderer;
 import com.arcanc.pulselib.content.renderer.modelData.DefaultBlockModelData;
 import com.arcanc.pulselib.content.renderer.modelData.PModelData;
@@ -40,7 +40,7 @@ public class BioForgeRenderer extends PBlockRenderer<BioForge>
         return animatable.getBlockState().getValue(BioForgeBlock.DOUBLE) ? DOUBLE_FORGE : FORGE;
     }
     
-    public static void registerTextures(final CustomEvents.PLibRegisterTextureEvent event)
+    public static void registerTextures(final PulseLibEvents.RegisterTextureEvent event)
     {
         event.addTextureLocation(TEXTURE);
     }