+++ src/main/java/com/arcanc/biomorphosis/content/block/block_entity/ber/BioSqueezerRenderer.java
@@ -14,7 +14,7 @@ import com.arcanc.biomorphosis.content.block.block_entity.BioSqueezer;
 import com.arcanc.biomorphosis.util.Database;
 import com.arcanc.biomorphosis.util.helper.FluidHelper;
 import com.arcanc.pulselib.content.animatable.PAnimationController;
-import com.arcanc.pulselib.content.event.CustomEvents;
+import com.arcanc.pulselib.content.event.PulseLibEvents;
 import com.arcanc.pulselib.content.model.baked.PBakedBone;
 import com.arcanc.pulselib.content.renderer.PBlockRenderer;
 import com.arcanc.pulselib.content.renderer.modelData.DefaultBlockModelData;
@@ -55,7 +55,7 @@ public class BioSqueezerRenderer extends PBlockRenderer<BioSqueezer>
 		poseStack.popPose();
 	}
 	
-	public static void registerTextures(final CustomEvents.PLibRegisterTextureEvent event)
+	public static void registerTextures(final PulseLibEvents.RegisterTextureEvent event)
 	{
 		event.addTextureLocation(TEXTURE);
 	}