+++ src/main/java/com/arcanc/biomorphosis/content/entity/renderer/MelonMawRenderer.java
@@ -12,7 +12,7 @@ package com.arcanc.biomorphosis.content.entity.renderer;
 
 import com.arcanc.biomorphosis.content.entity.MelonMaw;
 import com.arcanc.biomorphosis.util.Database;
-import com.arcanc.pulselib.content.event.CustomEvents;
+import com.arcanc.pulselib.content.event.PulseLibEvents;
 import com.arcanc.pulselib.content.renderer.PEntityRenderer;
 import com.arcanc.pulselib.content.renderer.modelData.DefaultEntityModelData;
 import com.arcanc.pulselib.util.PRenderTypes;
@@ -30,7 +30,7 @@ public class MelonMawRenderer extends PEntityRenderer<MelonMaw>
 				PRenderTypes.RenderTypeProvider :: trianglesSolid);
 	}
 	
-	public static void registerTextures(final CustomEvents.PLibRegisterTextureEvent event)
+	public static void registerTextures(final PulseLibEvents.RegisterTextureEvent event)
 	{
 		event.addTextureLocation(TEXTURE);
 	}