+++ src/main/java/com/arcanc/biomorphosis/content/worldgen/spawner/BioSpawner.java
@@ -89,7 +89,7 @@ public class BioSpawner
 		INFESTOR("infestor", 1, 2, 1),
 		KSIGG("ksigg", 1, 3, 1),
 		LARVA("larva", 1, 2, 1),
-		SWARMLING("swarmling", 1, 3, 1),
+		SWARMLING("swarmling", 1, 2, 1),
 		ZIRIS("ziris", 1, 0, 1),
 		MELONMAW("melonmaw", 1, 4, 1);