+++ changelog.md
@@ -1,2 +1,5 @@
 * Updated Infestor AI
 * Changed textures to smoother variant
+* added first version of organic armor
+* bumped pulselib version
+* nerfed infestor damage